+++ apps/chat-api/src/auth/keys/keys.service.spec.ts
