+++ apps/chat-api/src/auth/providers/provider-registry.service.spec.ts
